# outputs/api_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hospital DB Dashboard API",
    "description": "REST API for Hospital Database Dashboard - hidb10949",
    "version": "1.0.0",
    "contact": {
      "name": "Hospital DB Dashboard Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://hospital-db-dashboard.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/anomalies/{category}": {
      "get": {
        "tags": ["Anomalies"],
        "summary": "Get anomaly category details",
        "description": "Returns detailed list of anomalies for a specific category.",
        "operationId": "getAnomalyCategory",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "description": "Anomaly category",
            "schema": {
              "type": "string",
              "enum": ["missing_primary_keys", "missing_indexes", "large_tables_no_index", "orphan_tables", "naming_issues"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "severity",
            "in": "query",
            "description": "Filter by severity",
            "schema": {
              "type": "string",
              "enum": ["all", "critical", "warning", "info"],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnomalyCategoryResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "category": "missing_primary_keys",
                    "description": "Tables without primary key",
                    "total": 192,
                    "by_severity": {
                      "critical": 24,
                      "warning": 45,
                      "info": 123
                    },
                    "items": [
                      {
                        "table": "data_correct",
                        "row_count": 8437993,
                        "severity": "CRITICAL",
                        "module": "other",
                        "recommendation": "Add PRIMARY KEY on (hospcode, tablename)"
                      },
                      {
                        "table": "PERSON",
                        "row_count": 105707,
                        "severity": "CRITICAL",
                        "module": "patient",
                        "recommendation": "Add PRIMARY KEY on 'id' column"
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 50,
                      "total": 192,
                      "total_pages": 4
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tables": {
      "get": {
        "tags": ["Tables"],
        "summary": "Get all tables (paginated)",
        "description": "Returns paginated list of all database tables with metadata.",
        "operationId": "getAllTables",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "module",
            "in": "query",
            "description": "Filter by module name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search tables by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string",
              "enum": ["name", "row_count", "size_bytes"],
              "default": "name"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Sort order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablesListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search tables and modules",
        "description": "Search for tables and modules matching the query string.",
        "operationId": "searchDatabase",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Filter by type",
            "schema": {
              "type": "string",
              "enum": ["all", "tables", "modules"],
              "default": "all"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum results per category",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "query": "patient",
                    "results": {
                      "modules": [
                        {
                          "name": "patient",
                          "description": "ข้อมูลผู้ป่วยและบุคคล",
                          "tables_count": 130,
                          "relevance": 1.0
                        }
                      ],
                      "tables": [
                        {
                          "name": "PERSON",
                          "module": "patient",
                          "row_count": 105707,
                          "relevance": 0.95
                        },
                        {
                          "name": "pt",
                          "module": "patient",
                          "row_count": 181351,
                          "relevance": 0.85
                        },
                        {
                          "name": "pttype",
                          "module": "patient",
                          "row_count": 96,
                          "relevance": 0.80
                        }
                      ]
                    },
                    "total": {
                      "modules": 1,
                      "tables": 15
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid search query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "INVALID_QUERY",
                    "message": "Search query must be at least 1 character"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": ["success"]
      },
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "cached": {
            "type": "boolean"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string",
            "enum": ["static", "database"]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code",
                "example": "MODULE_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message",
                "example": "Module 'unknown' not found"
              },
              "details": {
                "type": "object",
                "description": "Additional error details"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["success", "error"]
      },
...
      "AnomalyCounts": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 527
          },
          "critical": {
            "type": "integer",
            "example": 24
          },
          "warning": {
            "type": "integer",
            "example": 152
          },
          "info": {
            "type": "integer",
            "example": 240
          }
        },
        "required": ["total", "critical", "warning", "info"]
      },
...
      "TableSummary": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "PERSON"
          },
          "row_count": {
            "type": "integer",
            "example": 105707
          },
          "size_bytes": {
            "type": "integer",
            "example": 18399232
          }
        },
        "required": ["name"]
      },
...
      "AnomalyCategoryResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "category": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "total": {
                    "type": "integer"
                  },
                  "by_severity": {
                    "$ref": "#/components/schemas/AnomalyCounts"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AnomalyItem"
                    }
                  },
                  "pagination": {
                    "$ref": "#/components/schemas/Pagination"
                  }
                },
                "required": ["category", "total", "items"]
              }
            },
            "required": ["data"]
          }
        ]
      },
      "AnomalyItem": {
        "type": "object",
        "properties": {
          "table": {
            "type": "string",
            "example": "PERSON"
          },
          "row_count": {
            "type": "integer",
            "example": 105707
          },
          "severity": {
            "type": "string",
            "enum": ["CRITICAL", "WARNING", "INFO"],
            "example": "CRITICAL"
          },
          "module": {
            "type": "string",
            "example": "patient"
          },
          "recommendation": {
            "type": "string",
            "example": "Add PRIMARY KEY on 'id' column"
          }
        },
        "required": ["table", "severity"]
      },
...
      "TablesListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "tables": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TableSummary"
                    }
                  },
                  "pagination": {
                    "$ref": "#/components/schemas/Pagination"
                  }
                },
                "required": ["tables", "pagination"]
              }
            },
            "required": ["data"]
          }
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 50
          },
          "total": {
            "type": "integer",
            "example": 847
          },
          "total_pages": {
            "type": "integer",
            "example": 17
          }
        },
        "required": ["page", "limit", "total", "total_pages"]
      },
      "SearchResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "example": "patient"
                  },
                  "results": {
                    "type": "object",
                    "properties": {
                      "modules": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/SearchResultModule"
                        }
                      },
                      "tables": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/SearchResultTable"
                        }
                      }
                    }
                  },
                  "total": {
                    "type": "object",
                    "properties": {
                      "modules": {
                        "type": "integer"
                      },
                      "tables": {
                        "type": "integer"
                      }
                    }
                  }
                },
                "required": ["query", "results"]
              }
            },
            "required": ["data"]
          }
        ]
      },
      "SearchResultModule": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "tables_count": {
            "type": "integer"
          },
          "relevance": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Search relevance score (0-1)"
          }
        },
        "required": ["name", "relevance"]
      },
      "SearchResultTable": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "module": {
            "type": "string"
          },
          "row_count": {
            "type": "integer"
          },
          "relevance": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": ["name", "relevance"]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Optional authentication for future API access (currently not required)"
      }
    }
  }
}